+++ src/bootstrap.rs
@@ -16,26 +16,30 @@ type TaskHandle = tokio::task::JoinHandle<Result<()>>;
 
 pub struct Application {
     config: Config,
-    router: Router,
+    router_fn: Option<Box<dyn FnOnce() -> Router + Send + Sync>>,
     pre_run_fn: Option<Box<dyn FnOnce() -> TaskHandle + Send + Sync>>,
 }
 
 impl Application {
-    pub fn default(config_path: &str, router: Router) -> Result<Self> {
+    pub fn default(config_path: &str) -> Result<Self> {
         let config = load_config(config_path).with_context(|| "configuration parsing failed")?;
-        Ok(Self {
-            config,
-            router,
-            pre_run_fn: None,
-        })
+        Ok(Self::new(config))
     }
 
-    pub fn new(config: Config, router: Router) -> Result<Self> {
-        Ok(Self {
+    pub fn new(config: Config) -> Self {
+        Self {
             config,
-            router,
+            router_fn: None,
             pre_run_fn: None,
-        })
+        }
+    }
+
+    pub fn with_router<F>(mut self, callback: F) -> Self
+    where
+        F: FnOnce() -> Router + Send + Sync + 'static,
+    {
+        self.router_fn = Some(Box::new(callback));
+        self
     }
 
     pub fn before_run<F>(mut self, callback: F) -> Self
@@ -62,7 +66,13 @@ impl Application {
         }
         let worker_guard =
             logger::init(&self.config.logger).with_context(|| "logger initialization failed")?;
-        general::serve(&self.config.general, self.router)
+        let router = self
+            .router_fn
+            .map(|callback| callback())
+            .unwrap_or_else(|| {
+                Router::new().route("/", axum::routing::get(|| async { "Hello, Axum-kit!" }))
+            });
+        general::serve(&self.config.general, router)
             .await
             .with_context(|| "service startup failed")?;